issue #8: fix tests after introducing html_content table
