+++ bin/psql.sh
@@ -1,3 +1,8 @@
 DIRNAME=`dirname $0`
 . $DIRNAME/lib.sh
-psql
+if [ -z "$1" ]; then
+  psql
+else
+  SQL_SCRIPT=$1
+  $PSQL_ADMIN -f "$SQL_SCRIPT"
+fi
