+++ sql/2023-08-09-issue8-html_content-table.sql
